+++ scripts/clean-dist.js
@@ -1,0 +1,8 @@
+/**
+ * Cleans dist directory before babel rewrite.
+ */
+
+const path = require( 'path' );
+const del = require( 'del' );
+
+del( path.resolve( 'dist' ) ).then( () => null );