+++ src/tasks/clean.js
@@ -1,7 +1,7 @@
 import gulp from 'gulp';
-import { tasks } from '../utils/get-config';
+import { cwd, tasks } from '../utils/config';
 import del from 'del';
-import TaskHelper from '../utils/TaskHelper';
+import TaskHelper from '../classes/TaskHelper';
 
 if ( tasks.clean ) {
 	const task = new TaskHelper( {
@@ -12,7 +12,7 @@ if ( tasks.clean ) {
 
 	gulp.task( task.name, done => {
 		if ( task.isValid() ) {
-			del( task.src ).then( () => done() );
+			del( task.src, { cwd } ).then( () => done() );
 		} else {
 			done();
 		}