+++ src/tasks/copy.js
@@ -1,8 +1,8 @@
 import gulp from 'gulp';
 import gulpIf from 'gulp-if';
 import cache from 'gulp-cached';
-import { tasks, isDev } from '../utils/get-config';
-import TaskHelper from '../utils/TaskHelper';
+import { tasks, isDev } from '../utils/config';
+import TaskHelper from '../classes/TaskHelper';
 
 if ( tasks.copy ) {
 	const task = new TaskHelper( {