+++ src/tasks/images.js
@@ -2,8 +2,8 @@ import gulp from 'gulp';
 import gulpIf from 'gulp-if';
 import cache from 'gulp-cached';
 import imagemin from 'gulp-imagemin';
-import { tasks, isDev } from '../utils/get-config';
-import TaskHelper from '../utils/TaskHelper';
+import { tasks, isDev } from '../utils/config';
+import TaskHelper from '../classes/TaskHelper';
 
 if ( tasks.images ) {
 	const task = new TaskHelper( {