+++ src/tasks/watch.js
@@ -1,22 +1,26 @@
 import gulp from 'gulp';
-import { tasks, cwd } from '../utils/get-config';
+import { tasks, cwd } from '../utils/config';
 import { join } from 'path';
 
 if ( tasks.watch && 0 < tasks.watch.length ) {
 	gulp.task( 'watch', () => {
 
 		// Omit some tasks, e.g. `js` is already watched by Webpack.
-		const ignoredTasks = [ 'js', 'js-lint', 'clean' ],
+		const ignoredTasks = [ 'clean' ],
 			filteredTasks = tasks.watch.filter( task => ! ignoredTasks.includes( task ) );
 
 		filteredTasks.forEach( taskSlug => {
 			const task = tasks[ taskSlug ];
 
-			if ( undefined === task.src ) {
+			if ( null === task || undefined === task || undefined === task.src ) {
 				return;
 			}
 
-			gulp.watch( join( cwd, task.src ), gulp.parallel( taskSlug ) );
+			if ( Array.isArray( task.src ) ) {
+				task.src.map( src => gulp.watch( join( cwd, src ), gulp.parallel( taskSlug ) ) );
+			} else {
+				gulp.watch( join( cwd, task.src ), gulp.parallel( taskSlug ) );
+			}
 		});
 	});
 }