+++ src/utils/config.js
@@ -1,17 +1,18 @@
 import fs from 'fs';
 import gutil from 'gulp-util';
 import yargs from 'yargs';
-import ConfigClass from './ConfigClass';
+import path from 'path';
+import Config from '../classes/Config';
 
-const configFile = './package.json';
+const configFile = path.resolve( './package.json' );
 let configJSON, config;
 
 if ( ! fs.existsSync( configFile ) ) {
 	gutil.log( gutil.colors.red( `Config file not found. Aborting...` ) );
-	process.exit( 1 );
+	process.exit( 1 ); // eslint-disable-line no-magic-numbers
 }
 
 configJSON = JSON.parse( fs.readFileSync( configFile ) );
-config = new ConfigClass( configJSON.workflows, yargs.argv );
+config = new Config( configJSON.workflows, yargs.argv );
 
 export const { cwd, env, isDev, isProd, isTest, tasks, workflow, workflowName } = config;