+++ src/utils/pre-check.js
@@ -1,8 +1,8 @@
 import gutil from 'gulp-util';
-import { workflowName, env } from './get-config';
+import { workflowName, env } from './config';
 
 export const preCheck = function() {
-	if ( undefined === workflowName ) {
+	if ( undefined !== workflowName ) {
 		gutil.log( `Using '${ gutil.colors.yellow( workflowName ) }' workflow...` );
 	}
 	if ( undefined !== env ) {