+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-# 0.1.0
+# 0.1.1 - 2017-09-05
 
--   Initial release.
+- Add support for custom tasks definition.
+- Move classes to the `classes` directory.
+- Rename classes so that they no longer have `Class` suffix.
+- Fix minor bugs.
+
+# 0.1.0 - 2017-07-26
+
+- Initial release.